pyproject.toml
[project]
name = "watvisionrevival"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.9",
    "azure-cognitiveservices-speech>=1.44.0",
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "azure-identity>=1.23.0",
    "dotenv>=0.9.9",
    "eventlet>=0.40.0",
    "flask>=3.1.1",
    "flask-socketio>=5.5.1",
    "kornia>=0.8.1",
    "matplotlib>=3.10.3",
    "mediapipe>=0.10.21",
    "openai[realtime]>=1.84.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-socketio>=5.13.0",
    "requests>=2.32.3",
    "torch>=2.7.0",
    "transformers>=4.52.3",
]
